+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 更改日志 | Change log
 
+## 2024-04-04
+
+- 实现关闭账户功能 | Implement close account feature
+
 ## 2024-04-03
 
 - 完善草稿管理功能的测试用例 | Improve test cases for draft management feature